Extract useHandleSwap() from Swap logic
